added test for pager grammar
Since the pager paper does not list all states I had
to add the renaming states by hand. For this reason I
can not guarantee that the graph is 100% correct, although
I am pretty confident it is.

Even though the pager paper lists the W-successor and V-successor
of state 2/8 and 3/9 as seperate, they should actually be merged
(they have the same core but only one item, which makes weakly always
succeed as there is nothing to compare).